requirements.txt
customtkinter==5.1.3
keyboard==0.13.5
Pillow==9.5.0
psutil==5.9.5
pywin32==306
